33-gruppen-anlegen-prio-1

+++ src/main/java/org/openlogisticsfoundation/ecmr/domain/services/UserService.java
@@ -11,30 +11,22 @@ package org.openlogisticsfoundation.ecmr.domain.services;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.openlogisticsfoundation.ecmr.domain.exceptions.EcmrNotFoundException;
 import org.openlogisticsfoundation.ecmr.domain.exceptions.GroupNotFoundException;
-import org.openlogisticsfoundation.ecmr.domain.exceptions.LocationNotFoundException;
+import org.openlogisticsfoundation.ecmr.domain.exceptions.NoPermissionException;
 import org.openlogisticsfoundation.ecmr.domain.exceptions.UserNotFoundException;
 import org.openlogisticsfoundation.ecmr.domain.exceptions.ValidationException;
 import org.openlogisticsfoundation.ecmr.domain.mappers.GroupPersistenceMapper;
-import org.openlogisticsfoundation.ecmr.domain.mappers.LocationPersistenceMapper;
 import org.openlogisticsfoundation.ecmr.domain.mappers.UserPersistenceMapper;
-import org.openlogisticsfoundation.ecmr.domain.models.CountryCode;
+import org.openlogisticsfoundation.ecmr.domain.models.AuthenticatedUser;
 import org.openlogisticsfoundation.ecmr.domain.models.Group;
-import org.openlogisticsfoundation.ecmr.domain.models.Location;
 import org.openlogisticsfoundation.ecmr.domain.models.User;
 import org.openlogisticsfoundation.ecmr.domain.models.commands.UserCommand;
-import org.openlogisticsfoundation.ecmr.persistence.entities.EcmrEntity;
 import org.openlogisticsfoundation.ecmr.persistence.entities.GroupEntity;
-import org.openlogisticsfoundation.ecmr.persistence.entities.LocationEntity;
 import org.openlogisticsfoundation.ecmr.persistence.entities.UserEntity;
 import org.openlogisticsfoundation.ecmr.persistence.entities.UserToGroupEntity;
-import org.openlogisticsfoundation.ecmr.persistence.entities.UserToLocationEntity;
 import org.openlogisticsfoundation.ecmr.persistence.repositories.GroupRepository;
-import org.openlogisticsfoundation.ecmr.persistence.repositories.LocationRepository;
 import org.openlogisticsfoundation.ecmr.persistence.repositories.UserRepository;
 import org.openlogisticsfoundation.ecmr.persistence.repositories.UserToGroupRepository;
-import org.openlogisticsfoundation.ecmr.persistence.repositories.UserToLocationRepository;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -48,31 +40,31 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final UserPersistenceMapper userPersistenceMapper;
-    private final LocationRepository locationRepository;
     private final GroupRepository groupRepository;
     private final UserToGroupRepository userToGroupRepository;
-    private final UserToLocationRepository userToLocationRepository;
     private final GroupPersistenceMapper groupPersistenceMapper;
-    private final LocationPersistenceMapper locationPersistenceMapper;
+    private final GroupService groupService;
 
     public List<User> getAllUsers() {
         return userRepository.findAll().stream().map(userPersistenceMapper::toUser).toList();
     }
 
-    public User createUser(@Valid UserCommand userCommand) throws ValidationException, LocationNotFoundException, GroupNotFoundException {
+    public User createUser(AuthenticatedUser authenticatedUser, @Valid UserCommand userCommand) throws ValidationException, GroupNotFoundException, NoPermissionException {
         UserEntity userEntity = userPersistenceMapper.toUserEntity(userCommand);
-        userEntity = userRepository.save(userEntity);
-        List<LocationEntity> locations = new ArrayList<>();
-        for (long locationId : userCommand.getLocationIds()) {
-            locations.add(locationRepository.findById(locationId).orElseThrow(() -> new LocationNotFoundException(locationId)));
-        }
         List<GroupEntity> groups = new ArrayList<>();
         for (long groupId : userCommand.getGroupIds()) {
             groups.add(groupRepository.findById(groupId).orElseThrow(() -> new GroupNotFoundException(groupId)));
         }
-        this.validateLocationsAndGroups(userCommand.getCountry(), locations, groups);
 
-        addGroupsAndLocationsToUser(userEntity, groups, locations);
+        if (!groupService.areAllGroupIdsPartOfUsersGroup(authenticatedUser, userCommand.getGroupIds())) {
+            throw new NoPermissionException("At least one group without permission");
+        }
+
+        userEntity = userRepository.save(userEntity);
+        for (GroupEntity groupEntity : groups) {
+            UserToGroupEntity userToGroupEntity = new UserToGroupEntity(userEntity, groupEntity);
+            this.userToGroupRepository.save(userToGroupEntity);
+        }
         return userPersistenceMapper.toUser(userEntity);
     }
 
@@ -80,35 +72,28 @@ public class UserService {
         return userToGroupRepository.findGroupsByUserId(userId).stream().map(groupPersistenceMapper::toGroup).toList();
     }
 
-    public List<Location> getLocationsByUserId(Long userId) {
-        return userToLocationRepository.findLocationsByUserId(userId).stream().map(locationPersistenceMapper::toLocation).toList();
-    }
-
     @Transactional
-    public User updateUser(long userId, UserCommand userCommand) throws UserNotFoundException, GroupNotFoundException, LocationNotFoundException, ValidationException {
+    public User updateUser(AuthenticatedUser authenticatedUser, long userId, UserCommand userCommand) throws UserNotFoundException, GroupNotFoundException, ValidationException, NoPermissionException {
         UserEntity userEntity = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
         userEntity = userPersistenceMapper.toUserEntity(userEntity, userCommand);
 
-        // Get the current groups and locations of the user
-        List<GroupEntity> groups = userToGroupRepository.findGroupsByUserId(userId);
-        List<LocationEntity> locations = userToLocationRepository.findLocationsByUserId(userId);
+        if (!groupService.areAllGroupIdsPartOfUsersGroup(authenticatedUser, userCommand.getGroupIds())) {
+            throw new NoPermissionException("At least one group without permission");
+        }
 
-        // Get a list of groups and locations that need to be deleted
+        // Get the current groups of the user
+        List<GroupEntity> groups = userToGroupRepository.findGroupsByUserId(userId);
+
+        // Get a list of groups that need to be deleted
         List<GroupEntity> groupsToDelete = groups.stream()
                 .filter(group -> !userCommand.getGroupIds().contains(group.getId())).toList();
-        List<LocationEntity> locationsToDelete = locations.stream()
-                .filter(location -> !userCommand.getLocationIds().contains(location.getId())).toList();
 
-        // Get a list of group and location ids that need to be added
+        // Get a list of groups ids that need to be added
         List<Long> currentGroupIds = groups.stream().map(GroupEntity::getId).toList();
         List<Long> groupIdsToAdd = userCommand.getGroupIds().stream().filter(groupId -> !currentGroupIds.contains(groupId)).toList();
 
-        List<Long> currentLocationIds = locations.stream().map(LocationEntity::getId).toList();
-        List<Long> locationIdsToAdd = userCommand.getLocationIds().stream().filter(locationId -> !currentLocationIds.contains(locationId)).toList();
-
-        // Remove old groups and locations
+        // Remove old groups
         groups.removeAll(groupsToDelete);
-        locations.removeAll(locationsToDelete);
 
         // Add new groups to two lists. One that contains all users groups (for validation) and one that contains only the new ones that have to be added to the database
         List<GroupEntity> groupsToAdd = new ArrayList<>();
@@ -118,46 +103,22 @@ public class UserService {
             groupsToAdd.add(group);
         }
 
-        // Add new locations to two lists. One that contains all users locations (for validation) and one that contains only the new ones that have to be added to the database
-        List<LocationEntity> locationsToAdd = new ArrayList<>();
-        for (long locationId : locationIdsToAdd) {
-            LocationEntity locationEntity = locationRepository.findById(locationId).orElseThrow(() -> new LocationNotFoundException(locationId));
-            locations.add(locationEntity);
-            locationsToAdd.add(locationEntity);
+        if (userEntity.getId() == authenticatedUser.getUser().getId() && (!groupsToAdd.isEmpty() || !groupsToDelete.isEmpty())) {
+            throw new NoPermissionException("A user can't change his own groups");
         }
 
-        // Check locations for correct countrycode and check if all groups are part of the locations
-        this.validateLocationsAndGroups(userEntity.getCountry(), locations, groups);
+        userEntity = userRepository.save(userEntity);
 
         // Delete old groups from database
         groupsToDelete.forEach(group -> userToGroupRepository.deleteByUserIdAndGroupId(userId, group.getId()));
-        locationsToDelete.forEach(location -> userToLocationRepository.deleteByUserIdAndLocationId(userId, location.getId()));
 
-        // Add new groups and locations to database
-        addGroupsAndLocationsToUser(userEntity, groupsToAdd, locationsToAdd);
-
-        return userPersistenceMapper.toUser(userEntity);
-    }
-
-    private void addGroupsAndLocationsToUser(UserEntity userEntity, List<GroupEntity> groups, List<LocationEntity> locations) {
-        for (GroupEntity groupEntity : groups) {
+        // Add new groups to database
+        for (GroupEntity groupEntity : groupsToAdd) {
             UserToGroupEntity userToGroupEntity = new UserToGroupEntity(userEntity, groupEntity);
             this.userToGroupRepository.save(userToGroupEntity);
         }
 
-        for (LocationEntity locationEntity : locations) {
-            UserToLocationEntity userToLocationEntity = new UserToLocationEntity(userEntity, locationEntity);
-            this.userToLocationRepository.save(userToLocationEntity);
-        }
-    }
-
-    private void validateLocationsAndGroups(CountryCode userCountry, List<LocationEntity> locationEntities, List<GroupEntity> groupEntities) throws ValidationException {
-        if (locationEntities.stream().anyMatch(l -> l.getCountryCode() != userCountry)) {
-            throw new ValidationException("At least one location has not the same country code as the user");
-        }
-        if (groupEntities.stream().anyMatch(g -> !locationEntities.contains(g.getLocation()))) {
-            throw new ValidationException("At least one group is not part of the users locations");
-        }
+        return userPersistenceMapper.toUser(userEntity);
     }
 
     public List<User> getUsersByGroupId(long groupId) {
